Se realiza el dashboard de articulos
