+++ src/main/java/co/edu/ufps/controller/UserController.java
@@ -1,93 +1,77 @@
 package co.edu.ufps.controller;
 
-import java.util.List;
-import java.util.Optional;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpSession;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import co.edu.ufps.model.User;
-import co.edu.ufps.repository.UserRepository;
+import co.edu.ufps.services.UserService;
 
-@RestController
-@RequestMapping("/usuario")
+@Controller
 public class UserController {
-	
+
 	@Autowired
-	UserRepository userRepository;
-	
-	@GetMapping("/status")
-	public String status() {
-		return "ok";
+	private UserService userService;
+
+	@GetMapping("/")
+	public String login(Model model) {
+		return "index";
 	}
-	
-	@GetMapping
-	public List <User> getAll() {
-		return (List<User>) userRepository.findAll();
-	}
-	
-	@GetMapping("/{id}")
-	public User getObject(@PathVariable Integer id) {
-		
-		Optional user = userRepository.findById(id);
-		if(user.isPresent()) {
-			User u = (User) user.get();
-			return u;
+
+	@PostMapping("/login")
+	public String loginUser(RedirectAttributes ra , String email, String clave, HttpServletRequest request, HttpSession session, Model model) {
+
+		User user = userService.select(email, clave);
+		if (user != null) {
+			request.getSession().setAttribute("user_id", user.getId());
+			request.getSession().setAttribute("pais", user.getPais());
+			return "menu-principal";
+		} else {
+			ra.addFlashAttribute("loginError", "Usuario o contraseña incorrecto");
+			return "redirect:/";
 		}
-		return null;
 	}
-	
-	@PostMapping
-	public User postObject(@RequestBody User userIn) {
-		
-		User entity = new User();
-		entity.setNombre(userIn.getNombre());
-		entity.setEmail(userIn.getEmail());
-		entity.setClave(userIn.getClave());
-		entity.setPais(userIn.getPais());
-		entity.setEntidad(userIn.getEntidad());
-		entity.setEstado(userIn.getEstado());
-		
-		userRepository.save(entity);
-		return entity;
+
+
+	@GetMapping("/logout")
+	public String cerrarSesion(HttpSession session){
+		session.invalidate();
+		return "redirect:/";
 	}
-	
-	@PutMapping("/{id}")
-	public User putObject( @PathVariable Integer id, @RequestBody User userIn) {
-		
-		Optional user = userRepository.findById(id);
-		if(user.isPresent()) {
-		User u = (User) user.get();
-		User entity = u;
-		entity.setNombre(userIn.getNombre());
-		entity.setEmail(userIn.getEmail());
-		entity.setClave(userIn.getClave());
-		entity.setPais(userIn.getPais());
-		entity.setEntidad(userIn.getEntidad());
-		entity.setEstado(userIn.getEstado());
-		userRepository.save(entity);
-		return entity;
-		}
-		
-		return null;
+
+	@RequestMapping("/list")
+	public String listUsers(Model model) {
+		model.addAttribute("list", userService.getAll());
+		return "mostrar";
 	}
-	
-	@DeleteMapping("/{id}")
-	public User deleteObject(@PathVariable Integer id) {
-		
-		Optional user = userRepository.findById(id);
-		if(user.isPresent()) {
-			User u = (User) user.get();
-			userRepository.delete(u);
-			return u;
-		}
-		return null;
+
+	@GetMapping("/show")
+	public String showNewForm(Model model) {
+		return "register";
+	}
+
+	@PostMapping("/save")
+	public String register(User user, Model model) {
+		userService.save(user);
+		return "redirect:/list";
+	}
+
+	@GetMapping("/delete/{id}")
+	public String delete(@PathVariable Integer id, Model model) {
+		userService.delete(id);
+		return "redirect:/list";
+	}
+
+	@GetMapping("/menu")
+	public String verMenu(){
+		return "menu-principal";
 	}
 }
