+++ src/main/java/co/edu/ufps/service/impl/UserServiceImpl.java
@@ -21,14 +21,11 @@ public class UserServiceImpl extends GenericServiceImpl<User, Integer> implement
 		return userRepository;
 	}
 
+	@Override
 	public User select(String email, String clave) {
 		return userRepository.findByEmailAndClave(email, clave);
 	}
 
-
-
-
-	
 
 
 }
