+++ src/main/resources/templates/menu-principal.html
@@ -17,7 +17,7 @@
 
     <div class="flex justify-evenly">
         <figure class="text-center">
-            <a class="mx-auto flex items-center justify-center w-32 h-32 rounded-md bg-red-500" href="/view-articles">
+            <a class="mx-auto flex items-center justify-center w-32 h-32 rounded-md bg-red-500" href="/articulo/list">
                 <i class="fa-solid text-white text-6xl fa-newspaper"></i>
             </a>
             <p class="font-bold">Articulos</p>
